+++ src/com/puzzle/cost/CostFunction.java
@@ -2,8 +2,8 @@ package com.puzzle.cost;
 
 public interface CostFunction {
     int calculateCost(int parentCost, int sequence);
-    int calculateH (int sequence);
-    int calculateF(int parentCost, int sequence);
+    float calculateH (int sequence);
+    float calculateF(int parentCost, int sequence);
 
 }
